+++ app/views/comparisons/gas_targets/_table.html.erb
@@ -1,22 +1,1 @@
-<%= component 'comparison_table',
-              report: report, advice_page: advice_page, table_name: table_name, index_params: index_params,
-              headers: headers, colgroups: colgroups, advice_page_tab: advice_page_tab do |c| %>
-  <% @results.each do |result| %>
-    <% c.with_row do |r| %>
-      <% r.with_school school: result.school %>
-      <% r.with_var val: result.current_year_percent_of_target_relative, unit: :relative_percent, change: true %>
-      <% r.with_var val: result.current_year_unscaled_percent_of_target_relative, unit: :relative_percent, change: true %>
-      <% r.with_var val: result.current_year_kwh, unit: :kwh %>
-      <% r.with_var val: result.current_year_target_kwh, unit: :kwh %>
-      <% r.with_var val: result.unscaled_target_kwh_to_date, unit: :kwh %>
-      <% r.with_var val: result.tracking_start_date, unit: :date %>
-    <% end %>
-  <% end %>
-  <% c.with_footer do %>
-    <tr>
-      <td colspan="<%= headers.count %>">
-        <a name="footnotes"><strong><%= t('analytics.benchmarking.content.footnotes.notes') %></strong></a>
-        <%= t('analytics.benchmarking.configuration.column_heading_explanation.last_year_definition_html') %>
-      </td>
-    </tr>
-  <% end %><% end %>
+<%= render partial: 'comparisons/shared/targets/table', locals: local_assigns %>
